third_party/WebKit/LayoutTests/http/tests/htmlimports/import-custom-element-order.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="resources/import-custom-element-helper.js"></script>
<script>


var Hello = registerTestingCustomElement('x-hello');
var numberOfLinks = 2;
var t1 = async_test('Custom elements in async imports wait preceeding sync import before upgraded.');
var t2 = async_test('Custom elements in async imports wait preceeding sync import before resolved.');

var latch = new ImportTestLatch(function() {
    window.setTimeout(function() {
        t1.step(function() {
            assert_array_equals(['hello-slow', 'hello-2', 'hello-1'], Hello.ids);
            t1.done();
        });

        t2.step(function() {
            var Bye = registerTestingCustomElement('x-bye');
            assert_array_equals(['bye-slow', 'bye-2', 'bye-1'], Bye.ids);
            t2.done();
        });
    }, 0);
}, numberOfLinks);
</script>
</head>
<body>
<link rel=import href="resources/import-slow-custom-element-hello.cgi" onload="latch.loaded()">
<link rel=import async href="resources/import-custom-element-hello-1.html" onload="latch.loaded()">
<x-hello id="hello-2"></x-hello>
<x-bye id="bye-2"></x-bye>
</body>
</html>
